+++ tests/testthat/test-merge_taxa.R
@@ -5,10 +5,14 @@
 
 context("merge taxon concepts")
 
-tmp <- subset(Easplist, TaxonConceptID %in% c(206,197), slot="relations")
-
 test_that("taxa are merged", {
-			expect_equal(nrow(tmp@taxonRelations), 2)
-			expect_equal(nrow(merge_taxa(tmp, c(206, 197))@taxonRelations), 1)
+			expect_equal(nrow(Easplist@taxonRelations) >
+							nrow(merge_taxa(Easplist,
+											c(206, 197))@taxonRelations),
+					TRUE)
+			expect_is(merge_taxa(Easplist, level="species"), "taxlist")
+			## expect_equal(summary(merge_taxa(Easplist,
+			##                         level="species")@taxonRelations$Level)["subspecies"],
+			##         0)
 		}
 )